+++ src/idpconfgen/components/bgeo_strategies/exact/__init__.py
@@ -1,0 +1,3 @@
+"""Module to build backbone geometry angles with exact values."""
+
+name = "exact"